requirements.txt
olca-ipc==2.4.0
psycopg2-binary==2.9.10

